# Access the URL rewriter meta data structure using a pointer