Need to do some more tests on this, the idea is to avoid changing other parts of the code like this: https://github.com/TykTechnologies/tyk/commit/c2f80ff6bac2f2c7c92c559be3ae02a452a5c882#diff-e174c6f6100dfee8c74de8323d464e36

With this PR, the regular expressions used in the URL rewriter are compiled only once, following the idea of #1629, #1624. This is part of #1662.