@matiasinsaurralde 2018-05-04
Any feedback on this is appreciated, not sure it will affect other existing functionality, tests look ok