@dencoded 2018-05-04
avoid doing compile is cool. I was also thinking about creating some kind of internal service/wrapper arround `regexp` functions where we can cache results of all match/find/submatch like operations and avoid doing them every time, say some `map[string]bool` where key is just regexp+tested_string

say we have 100 requests per sec of endpoint calls `GET /abcxyz` and it matches regexp `abc.*` - it is not efficient to do 100 `regexp.Match` call per sec knowing the result is the same - it does lots of CPU and memory allocations under the hood. maybe this kind of caching is already done on package `"regexp"` level, needs to be proved 1st if we wanna go with that approach.