@matiasinsaurralde 2018-05-11
`benchcmp` output (will refactor the benchmark on a different PR)
```
% benchcmp old new
benchmark               old ns/op     new ns/op     delta
BenchmarkRewriter-8     197470        173320        -12.23%

benchmark               old allocs     new allocs     delta
BenchmarkRewriter-8     501            501            +0.00%

benchmark               old bytes     new bytes     delta
BenchmarkRewriter-8     378877        378876        -0.00%
```